Move broadway related tests to separate namespace
