+++ tests/Broadway/CodeGenerator/sample-files/broadway-code/broadway-event.output
