+++ tests/Broadway/CodeGenerator/sample-files/broadway-code/read/elasticsearch/repository.output
