+++ tests/Broadway/CodeGenerator/sample-files/broadway-code/uuid-identifier.output
