+++ tests/Broadway/CodeGenerator/sample-files/broadway-spec/broadway-command-spec.output
