+++ tests/Broadway/CodeGenerator/sample-files/broadway-spec/broadway-event-spec.output
