+++ tests/Broadway/CodeGenerator/sample-files/broadway-spec/broadway-model-repository-spec.output
