+++ tests/Broadway/CodeGenerator/sample-files/broadway-spec/broadway-model-spec.output
