+++ tests/Broadway/CodeGenerator/sample-files/broadway-spec/read/elasticsearch/entity-spec.output
