+++ tests/Broadway/CodeGenerator/sample-files/broadway-spec/read/elasticsearch/projector-spec.output
