+++ tests/Broadway/CodeGenerator/sample-files/broadway-spec/read/elasticsearch/repository-spec.output
