+++ tests/Broadway/CodeGenerator/sample-files/broadway-spec/uuid-identifier-spec.output
